release: v0.0.13 — IO format readers, smiles into molrs-io, Rust 1.95 (#11)
## Highlights since v0.0.12

- **io**: new readers for DCD trajectory, CIF, GRO, MOL2, POSCAR, SDF.
- **smiles**: standalone `molcrafts-molrs-smiles` crate folded into `molrs-io::smiles` behind a `smiles` feature; façade re-export preserved.
- **python**: rewired bindings to the new `molrs-io::smiles` path; fixes a build break on `dev` introduced when the standalone crate was removed.
- **wasm/capi/python**: Cargo manifests refreshed for the reshuffled `molrs-io` module layout.
- **lammps_dump**: minor reader tweaks + data-file integration test.
- **toolchain**: `rust-toolchain.toml` pins `channel = "stable"`; CI already used latest stable, so local now matches.
- **clippy**: fixes for new Rust 1.95 lints (`useless_conversion` in `molrs-io/src/cif.rs`, `explicit_auto_deref` in `molrs-core`, plus `#[allow(too_many_arguments)]` on the WASM `Grid::new` JS-binding constructor).
- **docs/skills**: align `CLAUDE.md` and the `molrs-documenter` agent with the reorganized skills, `argument-hint` frontmatter, standardized `Output` section.
- **release**: bumps every crate manifest from 0.0.12 → 0.0.13.

+++ molrs-python/tests/test_region.py
@@ -5,7 +5,7 @@ import molrs
 
 class TestSphere:
     def test_contains(self):
-        s = molrs.Sphere(np.array([0.0, 0.0, 0.0], dtype=np.float32), 2.0)
+        s = molrs.Sphere(np.array([0.0, 0.0, 0.0], dtype=np.float64), 2.0)
         pts = np.array(
             [
                 [0.0, 0.0, 0.0],  # center
@@ -13,14 +13,14 @@ class TestSphere:
                 [2.0, 0.0, 0.0],  # on surface
                 [2.1, 0.0, 0.0],  # outside
             ],
-            dtype=np.float32,
+            dtype=np.float64,
         )
         mask = s.contains(pts)
         assert mask[0] and mask[1] and mask[2]
         assert not mask[3]
 
     def test_bounds(self):
-        s = molrs.Sphere(np.array([1.0, 2.0, 3.0], dtype=np.float32), 2.0)
+        s = molrs.Sphere(np.array([1.0, 2.0, 3.0], dtype=np.float64), 2.0)
         b = s.bounds()
         assert b.shape == (3, 2)
         np.testing.assert_allclose(b[:, 0], [-1.0, 0.0, 1.0], atol=1e-5)
@@ -28,15 +28,15 @@ class TestSphere:
 
     def test_bad_center_shape(self):
         with pytest.raises(ValueError, match="length 3"):
-            molrs.Sphere(np.array([0.0, 0.0], dtype=np.float32), 1.0)
+            molrs.Sphere(np.array([0.0, 0.0], dtype=np.float64), 1.0)
 
     def test_bad_points_shape(self):
-        s = molrs.Sphere(np.array([0.0, 0.0, 0.0], dtype=np.float32), 1.0)
+        s = molrs.Sphere(np.array([0.0, 0.0, 0.0], dtype=np.float64), 1.0)
         with pytest.raises(ValueError, match="N, 3"):
-            s.contains(np.ones((3, 2), dtype=np.float32))
+            s.contains(np.ones((3, 2), dtype=np.float64))
 
     def test_repr(self):
-        s = molrs.Sphere(np.array([1.0, 2.0, 3.0], dtype=np.float32), 5.0)
+        s = molrs.Sphere(np.array([1.0, 2.0, 3.0], dtype=np.float64), 5.0)
         r = repr(s)
         assert "Sphere" in r
         assert "5.00" in r
@@ -45,7 +45,7 @@ class TestSphere:
 class TestHollowSphere:
     def test_contains(self):
         hs = molrs.HollowSphere(
-            np.array([0.0, 0.0, 0.0], dtype=np.float32), 2.0, 5.0
+            np.array([0.0, 0.0, 0.0], dtype=np.float64), 2.0, 5.0
         )
         pts = np.array(
             [
@@ -55,7 +55,7 @@ class TestHollowSphere:
                 [5.0, 0.0, 0.0],  # on outer surface
                 [5.1, 0.0, 0.0],  # outside
             ],
-            dtype=np.float32,
+            dtype=np.float64,
         )
         mask = hs.contains(pts)
         assert not mask[0] and not mask[1]
@@ -64,7 +64,7 @@ class TestHollowSphere:
 
     def test_bounds(self):
         hs = molrs.HollowSphere(
-            np.array([0.0, 0.0, 0.0], dtype=np.float32), 2.0, 5.0
+            np.array([0.0, 0.0, 0.0], dtype=np.float64), 2.0, 5.0
         )
         b = hs.bounds()
         assert b.shape == (3, 2)
@@ -72,7 +72,7 @@ class TestHollowSphere:
 
     def test_repr(self):
         hs = molrs.HollowSphere(
-            np.array([0.0, 0.0, 0.0], dtype=np.float32), 2.0, 5.0
+            np.array([0.0, 0.0, 0.0], dtype=np.float64), 2.0, 5.0
         )
         assert "HollowSphere" in repr(hs)
         assert "2.00" in repr(hs)
@@ -80,8 +80,8 @@ class TestHollowSphere:
 
 class TestRegionAnd:
     def test_sphere_and_sphere(self):
-        s1 = molrs.Sphere(np.array([0.0, 0.0, 0.0], dtype=np.float32), 3.0)
-        s2 = molrs.Sphere(np.array([2.0, 0.0, 0.0], dtype=np.float32), 3.0)
+        s1 = molrs.Sphere(np.array([0.0, 0.0, 0.0], dtype=np.float64), 3.0)
+        s2 = molrs.Sphere(np.array([2.0, 0.0, 0.0], dtype=np.float64), 3.0)
         intersection = s1 & s2
 
         pts = np.array(
@@ -90,24 +90,24 @@ class TestRegionAnd:
                 [-2.5, 0.0, 0.0],  # inside s1 only
                 [4.5, 0.0, 0.0],  # inside s2 only
             ],
-            dtype=np.float32,
+            dtype=np.float64,
         )
         mask = intersection.contains(pts)
         assert mask[0] and not mask[1] and not mask[2]
 
     def test_hollow_and_sphere(self):
         hs = molrs.HollowSphere(
-            np.array([0.0, 0.0, 0.0], dtype=np.float32), 1.0, 5.0
+            np.array([0.0, 0.0, 0.0], dtype=np.float64), 1.0, 5.0
         )
-        s = molrs.Sphere(np.array([3.0, 0.0, 0.0], dtype=np.float32), 3.0)
+        s = molrs.Sphere(np.array([3.0, 0.0, 0.0], dtype=np.float64), 3.0)
         result = hs & s
         assert isinstance(result, molrs.Region)
 
 
 class TestRegionOr:
     def test_sphere_or_sphere(self):
-        s1 = molrs.Sphere(np.array([0.0, 0.0, 0.0], dtype=np.float32), 1.0)
-        s2 = molrs.Sphere(np.array([5.0, 0.0, 0.0], dtype=np.float32), 1.0)
+        s1 = molrs.Sphere(np.array([0.0, 0.0, 0.0], dtype=np.float64), 1.0)
+        s2 = molrs.Sphere(np.array([5.0, 0.0, 0.0], dtype=np.float64), 1.0)
         union = s1 | s2
 
         pts = np.array(
@@ -116,7 +116,7 @@ class TestRegionOr:
                 [5.0, 0.0, 0.0],  # inside s2
                 [2.5, 0.0, 0.0],  # outside both
             ],
-            dtype=np.float32,
+            dtype=np.float64,
         )
         mask = union.contains(pts)
         assert mask[0] and mask[1] and not mask[2]
@@ -124,7 +124,7 @@ class TestRegionOr:
 
 class TestRegionNot:
     def test_not_sphere(self):
-        s = molrs.Sphere(np.array([0.0, 0.0, 0.0], dtype=np.float32), 2.0)
+        s = molrs.Sphere(np.array([0.0, 0.0, 0.0], dtype=np.float64), 2.0)
         complement = ~s
 
         pts = np.array(
@@ -132,14 +132,14 @@ class TestRegionNot:
                 [0.0, 0.0, 0.0],  # inside sphere
                 [3.0, 0.0, 0.0],  # outside sphere
             ],
-            dtype=np.float32,
+            dtype=np.float64,
         )
         mask = complement.contains(pts)
         assert not mask[0] and mask[1]
 
     def test_not_hollow_sphere(self):
         hs = molrs.HollowSphere(
-            np.array([0.0, 0.0, 0.0], dtype=np.float32), 2.0, 5.0
+            np.array([0.0, 0.0, 0.0], dtype=np.float64), 2.0, 5.0
         )
         result = ~hs
         assert isinstance(result, molrs.Region)
@@ -147,8 +147,8 @@ class TestRegionNot:
 
 class TestRegionChaining:
     def test_shell_via_and_not(self):
-        outer = molrs.Sphere(np.array([0.0, 0.0, 0.0], dtype=np.float32), 5.0)
-        inner = molrs.Sphere(np.array([0.0, 0.0, 0.0], dtype=np.float32), 2.0)
+        outer = molrs.Sphere(np.array([0.0, 0.0, 0.0], dtype=np.float64), 5.0)
+        inner = molrs.Sphere(np.array([0.0, 0.0, 0.0], dtype=np.float64), 2.0)
         shell = outer & (~inner)
 
         pts = np.array(
@@ -157,32 +157,32 @@ class TestRegionChaining:
                 [3.0, 0.0, 0.0],  # in shell
                 [6.0, 0.0, 0.0],  # outside
             ],
-            dtype=np.float32,
+            dtype=np.float64,
         )
         mask = shell.contains(pts)
         assert not mask[0] and mask[1] and not mask[2]
 
     def test_composed_bounds(self):
-        s1 = molrs.Sphere(np.array([0.0, 0.0, 0.0], dtype=np.float32), 3.0)
-        s2 = molrs.Sphere(np.array([2.0, 0.0, 0.0], dtype=np.float32), 3.0)
+        s1 = molrs.Sphere(np.array([0.0, 0.0, 0.0], dtype=np.float64), 3.0)
+        s2 = molrs.Sphere(np.array([2.0, 0.0, 0.0], dtype=np.float64), 3.0)
         result = s1 & s2
         b = result.bounds()
         assert b.shape == (3, 2)
 
     def test_composed_repr(self):
-        s1 = molrs.Sphere(np.array([0.0, 0.0, 0.0], dtype=np.float32), 3.0)
-        s2 = molrs.Sphere(np.array([2.0, 0.0, 0.0], dtype=np.float32), 3.0)
+        s1 = molrs.Sphere(np.array([0.0, 0.0, 0.0], dtype=np.float64), 3.0)
+        s2 = molrs.Sphere(np.array([2.0, 0.0, 0.0], dtype=np.float64), 3.0)
         assert "composed" in repr(s1 & s2)
 
     def test_region_and_region(self):
-        s1 = molrs.Sphere(np.array([0.0, 0.0, 0.0], dtype=np.float32), 5.0)
-        s2 = molrs.Sphere(np.array([0.0, 0.0, 0.0], dtype=np.float32), 3.0)
+        s1 = molrs.Sphere(np.array([0.0, 0.0, 0.0], dtype=np.float64), 5.0)
+        s2 = molrs.Sphere(np.array([0.0, 0.0, 0.0], dtype=np.float64), 3.0)
         r1 = s1 & s2
         r2 = ~s2
         r3 = r1 | r2  # composed & composed
         assert isinstance(r3, molrs.Region)
 
     def test_type_error_on_bad_operand(self):
-        s = molrs.Sphere(np.array([0.0, 0.0, 0.0], dtype=np.float32), 1.0)
+        s = molrs.Sphere(np.array([0.0, 0.0, 0.0], dtype=np.float64), 1.0)
         with pytest.raises(TypeError):
             s & "not_a_region"
